requirements.txt
streamlit>=1.32.0
yfinance>=0.2.38
groq>=0.5.0
requests>=2.31.0
gtts>=2.5.0
moviepy>=1.0.3
Pillow>=10.0.0
numpy>=1.24.0
